+++ src/Driver/Driver.php
@@ -33,22 +33,13 @@ abstract class Driver implements EventListenerInterface
     protected $connection;
 
     /**
-     * Database configuration
-     * @since 2.1.0
-     * @var array
-     */
-    protected $config;
-
-    /**
      * Construct
      * @param \Cake\Datasource\ConnectionInterface $connection A connection object
-     * @uses $config
      * @uses $connection
      */
     public function __construct($connection)
     {
         $this->connection = $connection;
-        $this->config = $connection->config();
 
         //Attachs the object to the event manager
         $this->getEventManager()->on($this);
@@ -138,7 +129,7 @@ abstract class Driver implements EventListenerInterface
             $executable .= ' | ' . $this->getBinary($compression);
         }
 
-        $executable .= ' > ' . $filename;
+        $executable .= ' > ' . escapeshellarg($filename);
 
         if (Configure::read(DATABASE_BACKUP . '.redirectStderrToDevNull')) {
             $executable .= REDIRECT_TO_DEV_NULL;
@@ -157,6 +148,7 @@ abstract class Driver implements EventListenerInterface
     {
         $executable = $this->_importExecutable();
         $compression = $this->getCompression($filename);
+        $filename = escapeshellarg($filename);
 
         if ($compression) {
             $executable = sprintf('%s -dc %s | ', $this->getBinary($compression), $filename) . $executable;
@@ -200,6 +192,29 @@ abstract class Driver implements EventListenerInterface
         }
 
         return file_exists($filename);
+    }
+
+    /**
+     * Gets a config value or the whole configuration
+     * @param string $key Config key
+     * @return array|string|null Config value, `null` if the key doesn't exist
+     *  or all config values if no key was specified
+     * @since 2.2.1
+     * @uses $connection
+     */
+    final public function getConfig($key = null)
+    {
+        $config = $this->connection->config();
+
+        if (!$key) {
+            return $config;
+        }
+
+        if (!array_key_exists($key, $config)) {
+            return null;
+        }
+
+        return $config[$key];
     }
 
     /**